pyproject.toml
[project]
name = "llm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "openai>=2.3.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "pygithub>=2.8.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.57.0",
    "uvicorn>=0.37.0",
]
